Fix prop warnings along with related bugs
