+++ src/components/EditableList/Preview.js
@@ -1,40 +1,44 @@
-import React from 'react';
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Flipped } from 'react-flip-toolkit';
 import { Item } from '../OrderedList';
 
-const Preview = ({
-  onClickPrompt,
-  editField,
-  fieldId,
-  onDelete,
-  sortable,
-}) => {
-  // Switch out flipId when fields match to prevent removal
-  // from DOM but still allow FLIP to work
-  const flipId = editField === fieldId ? `_${fieldId}` : fieldId;
+class Preview extends Component {
+  render() {
+    const {
+      onClickPrompt,
+      editField,
+      fieldId,
+      onDelete,
+      sortable,
+    } = this.props;
 
-  const handleClick = () => onClickPrompt(fieldId);
+    // Switch out flipId when fields match to prevent removal
+    // from DOM but still allow FLIP to work
+    const flipId = editField === fieldId ? `_${fieldId}` : fieldId;
 
-  return (
-    <Flipped flipId={flipId}>
-      {flipProps => (
-        <Item
-          {...flipProps}
-          sortable={sortable}
-          onDelete={onDelete}
-          onClick={handleClick}
-        >
-          <Flipped inverseFlipId={fieldId} scale>
-            <div className="editable-list-preview">
-              {this.preview && this.preview()}
-            </div>
-          </Flipped>
-        </Item>
-      )}
-    </Flipped>
-  );
-};
+    const handleClick = () => onClickPrompt(fieldId);
+
+    return (
+      <Flipped flipId={flipId}>
+        {flipProps => (
+          <Item
+            {...flipProps}
+            sortable={sortable}
+            onDelete={onDelete}
+            onClick={handleClick}
+          >
+            <Flipped inverseFlipId={fieldId} scale>
+              <div className="editable-list-preview">
+                {this.preview && this.preview()}
+              </div>
+            </Flipped>
+          </Item>
+        )}
+      </Flipped>
+    );
+  }
+}
 
 Preview.propTypes = {
   fieldId: PropTypes.string.isRequired,
