+++ src/components/Grid/GridItem.js
@@ -43,11 +43,15 @@ const GridItem = ({
 
 GridItem.propTypes = {
   fields: PropTypes.object.isRequired,
-  editField: PropTypes.string.isRequired,
+  editField: PropTypes.string,
   onEditItem: PropTypes.func.isRequired,
   previewComponent: PropTypes.any.isRequired,
   index: PropTypes.number.isRequired,
   id: PropTypes.string.isRequired,
+};
+
+GridItem.defaultProps = {
+  editField: null,
 };
 
 export { GridItem };
